+++ lib/File.js
@@ -1,4 +1,4 @@
-const { ensureDir, resolveRootPath } = require("./utils");
+const { ensureDir, resolveRootPath } = require("../utils");
 const fs = require("fs");
 const chalk = require("chalk");
