+++ lib/Folder.js
