+++ lib/clearcra.defaultConfig.js
