+++ lib/fileContent.js
